Remove int_uint from net.

+++ components/net/storage_task.rs
@@ -86,17 +86,17 @@ impl StorageManager {
         }
     }
 
-    fn length(&self, sender: Sender<u32>, url: Url, storage_type: StorageType) {
+    fn length(&self, sender: Sender<usize>, url: Url, storage_type: StorageType) {
         let origin = self.get_origin_as_string(url);
         let data = self.select_data(storage_type);
-        sender.send(data.get(&origin).map_or(0u, |entry| entry.len()) as u32).unwrap();
+        sender.send(data.get(&origin).map_or(0, |entry| entry.len())).unwrap();
     }
 
     fn key(&self, sender: Sender<Option<DOMString>>, url: Url, storage_type: StorageType, index: u32) {
         let origin = self.get_origin_as_string(url);
         let data = self.select_data(storage_type);
         sender.send(data.get(&origin)
-                    .and_then(|entry| entry.keys().nth(index as uint))
+                    .and_then(|entry| entry.keys().nth(index as usize))
                     .map(|key| key.clone())).unwrap();
     }
